mod: ported direct map impl

+++ src/route/error.rs
@@ -1,10 +1,10 @@
 #[derive(Debug)]
 pub enum RouteError {
-    HashMapInsertError(String, u32)
+    HashMapInsertError(String, u8)
 }
 
-impl From<(&str, u32)> for RouteError {
-    fn from(value: (&str, u32)) -> Self {
+impl From<(&str, u8)> for RouteError {
+    fn from(value: (&str, u8)) -> Self {
         RouteError::HashMapInsertError(value.0.to_string(), value.1)
     }
 }
